+++ src/index.css
@@ -22,7 +22,7 @@ div {
     display: grid;
     justify-items: stretch;
     grid-template-columns: auto  70vh auto;
-    grid-template-rows: 20vh 20vh 20vh 20vh;
+    grid-template-rows: auto auto auto auto;
     grid-template-areas:
     "TimeBox TimeBox TimeBox"
     "Chat ActiveBox Music1"
@@ -67,6 +67,8 @@ container1{
 }
 .svg {
     cursor: pointer;
+    border-radius: 4px;
+    border: 2px solid #4CAF50;
 }
 .svg img {
     width: 35px;
@@ -83,4 +85,7 @@ container1{
 }
 .boxList div {
     padding: 2px;
+}
+.alice-carousel__wrapper{
+    text-align: center;
 }
